commander : fixed warnings int vs size_t comparison

+++ tests/unit/Net2Com-test.cpp
@@ -59,7 +59,7 @@ TEST(Net2ComTestGroup, WriteToDataPipe_ReturnsCorrectNumberOfBytesWritten)
 {
     char buffer[BUFFER_SIZE];
     const char* data = "SomeUsefullData";
-    int result;
+    size_t result;
 
     result = netman->WriteToDataPipe(data);
     commander->ReadFromDataPipe(buffer, BUFFER_SIZE);
@@ -72,7 +72,7 @@ TEST(Net2ComTestGroup, ReadFromDataPipe_ReturnsCorrectString)
 {
     char buffer[BUFFER_SIZE];
     const char* data = "myCommand";
-    int result;
+    size_t result;
 
     result = commander->WriteToDataPipe(data);
     netman->ReadFromDataPipe(buffer, BUFFER_SIZE);
@@ -129,7 +129,7 @@ TEST(Net2ComTestGroup, ReadFromInfoPipe_TryOverFlowBuffer_ReadUntilBufferIsFull)
     const int TO_SMALL = 5;
     char buffer[BUF_SIZE] = {'\0'};
     const char* data = "myCommand";
-    int result;
+    size_t result;
 
     result = commander->WriteToDataPipe(data);
     CHECK_EQUAL(strlen(data) + NULL_CHAR_LENGTH, result);
